# back_end/v2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "没有描述",
        "title": "超市进销存系统--数据库课设接口文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/detail/delete": {
            "post": {
                "description": "删除订单详情，返回外键检查结果和删除结果，如果有外键依赖则会提示",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "detail"
                ],
                "summary": "删除订单详情",
                "parameters": [
                    {
                        "description": "Detail Request",
                        "name": "detail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Detail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/detail/insert": {
            "post": {
                "description": "批量导入订单详情，并且返回成功失败列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "detail"
                ],
                "summary": "批量导入订单详情",
                "parameters": [
                    {
                        "description": "Detail Request",
                        "name": "list",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DetailList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DetailInsertResponse"
                        }
                    }
                }
            }
        },
...
        "/detail/update": {
            "post": {
                "description": "更新产品，除了id之外都能改，所有不会有外键冲突",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "detail"
                ],
                "summary": "更新订单详情",
                "parameters": [
                    {
                        "description": "Detail Request",
                        "name": "detail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Detail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/employee/delete": {
            "post": {
                "description": "验证账号密码，然后删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功或失败",
                        "schema": {
                            "type": "msg"
                        }
                    }
                }
            }
        },
        "/employee/deleteByAdmin": {
            "post": {
                "description": "需要操作者id，操作对象id，如果操作者不存在，操作对象不存在，操作权限不足都会返回对应错误",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "管理员删除其他用户",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteEmployeeByAdminRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/employee/insert": {
            "post": {
                "description": "批量导入用户，并且返回成功失败列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "批量导入用户",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "list",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EmployeeList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EmployeeInsertResult"
                        }
                    }
                }
            }
        },
        "/employee/login": {
            "post": {
                "description": "验证账号密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EmployeeLoginResponse"
                        }
                    }
                }
            }
        },
        "/employee/update": {
            "post": {
                "description": "验证账号密码，然后把密码改为新密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "用户修改信息",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EmployeeUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功或失败",
                        "schema": {
                            "type": "msg"
                        }
                    }
                }
            }
        },
...
        "/order/delete": {
            "post": {
                "description": "删除订单，返回外键检查结果和删除结果，如果有外键依赖则会提示",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "删除订单",
                "parameters": [
                    {
                        "description": "Order Request",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/order/insert": {
            "post": {
                "description": "批量导入订单，并且返回成功失败列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "批量导入订单",
                "parameters": [
                    {
                        "description": "Order Request",
                        "name": "list",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OrderList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderInsertResponse"
                        }
                    }
                }
            }
        },
...
        "/order/update": {
            "post": {
                "description": "更新产品，除了id之外都能改，所有不会有外键冲突",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "更新订单",
                "parameters": [
                    {
                        "description": "Order Request",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.DeleteEmployeeByAdminRequest": {
            "type": "object",
            "properties": {
                "operator": {
                    "$ref": "#/definitions/models.Employee"
                },
                "target": {
                    "$ref": "#/definitions/models.Employee"
                }
            }
        },
        "dto.DetailInsertResponse": {
            "type": "object",
            "properties": {
                "fail_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Detail"
                    }
                },
                "success_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Detail"
                    }
                }
            }
        },
        "dto.DetailList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Detail"
                    }
                }
            }
        },
        "dto.EmployeeInsertResult": {
            "type": "object",
            "properties": {
                "fail_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Employee"
                    }
                },
                "success_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Employee"
                    }
                }
            }
        },
        "dto.EmployeeList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Employee"
                    }
                }
            }
        },
        "dto.EmployeeLoginResponse": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                },
                "tel": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "dto.EmployeeUpdateRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                },
                "tel": {
                    "type": "string"
                }
            }
        },
        "dto.OrderInsertResponse": {
            "type": "object",
            "properties": {
                "fail_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Order"
                    }
                },
                "success_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Order"
                    }
                }
            }
        },
        "dto.OrderList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Order"
                    }
                }
            }
        },
...
        "models.Detail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "list_id": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "number"
                }
            }
        },
        "models.Employee": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                },
                "tel": {
                    "type": "string"
                }
            }
        },
        "models.Order": {
            "type": "object",
            "properties": {
                "employee_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "total_price": {
                    "type": "number"
                }
            }
        },
...
    }
}